common: chores++
internal loader naming revisions for consistency
