+++ src/loaders/external_jpg/tvgJpgLoader.cpp
@@ -58,7 +58,7 @@ JpgLoader::~JpgLoader()
 bool JpgLoader::open(const char* path)
 {
 #ifdef THORVG_FILE_IO_SUPPORT
-    if (!(data = (unsigned char*) LoadModule::open(path, size))) return false;
+    if (!(data = (unsigned char*) Loader::open(path, size))) return false;
 
     int width, height, subSample, colorSpace;
     if (tjDecompressHeader3(jpegDecompressor, data, size, &width, &height, &subSample, &colorSpace) < 0) return false;
@@ -97,7 +97,7 @@ bool JpgLoader::open(const char* data, uint32_t size, TVG_UNUSED const char* rpa
 
 bool JpgLoader::read()
 {
-    if (!LoadModule::read()) return true;
+    if (!Loader::read()) return true;
 
     if (w == 0 || h == 0) return false;
