+++ src/loaders/external_webp/tvgWebpLoader.cpp
@@ -68,7 +68,7 @@ WebpLoader::~WebpLoader()
 bool WebpLoader::open(const char* path)
 {
 #ifdef THORVG_FILE_IO_SUPPORT
-    if (!(data = (unsigned char*)LoadModule::open(path, size))) return false;
+    if (!(data = (unsigned char*)Loader::open(path, size))) return false;
 
     int width, height;
     if (!WebPGetInfo(data, size, &width, &height)) return false;
@@ -107,7 +107,7 @@ bool WebpLoader::open(const char* data, uint32_t size, TVG_UNUSED const char* rp
 
 bool WebpLoader::read()
 {
-    if (!LoadModule::read()) return true;
+    if (!Loader::read()) return true;
 
     if (!data || w == 0 || h == 0) return false;
