+++ src/loaders/lottie/tvgLottieBuilder.cpp
@@ -24,7 +24,7 @@
 #include "tvgCommon.h"
 #include "tvgMath.h"
 #include "tvgScene.h"
-#include "tvgLoadModule.h"
+#include "tvgLoader.h"
 #include "tvgLottieModel.h"
 #include "tvgLottieBuilder.h"
 #include "tvgLottieExpressions.h"
