+++ src/loaders/png/tvgPngLoader.cpp
@@ -71,7 +71,7 @@ PngLoader::~PngLoader()
 bool PngLoader::open(const char* path)
 {
 #ifdef THORVG_FILE_IO_SUPPORT
-    if (!(data = (unsigned char*)LoadModule::open(path, size))) return false;
+    if (!(data = (unsigned char*)Loader::open(path, size))) return false;
 
     lodepng_state_init(&state);
 
@@ -114,7 +114,7 @@ bool PngLoader::read()
 {
     if (!data || w == 0 || h == 0) return false;
 
-    if (!LoadModule::read()) return true;
+    if (!Loader::read()) return true;
 
     TaskScheduler::request(this);
