+++ src/renderer/tvgInitializer.cpp
@@ -22,7 +22,7 @@
 
 #include "tvgCommon.h"
 #include "tvgTaskScheduler.h"
-#include "tvgLoader.h"
+#include "tvgLoaderMgr.h"
 
 #ifdef THORVG_SW_RASTER_SUPPORT
     #include "tvgSwRenderer.h"
