+++ src/renderer/tvgLoader.h
@@ -23,23 +23,153 @@
 #ifndef _TVG_LOADER_H_
 #define _TVG_LOADER_H_
 
-#include "tvgLoadModule.h"
+#include <atomic>
+#include "tvgCommon.h"
+#include "tvgRender.h"
+#include "tvgInlist.h"
 
 namespace tvg
 {
 
-struct LoaderMgr
+struct AssetResolver
 {
-    static bool init();
-    static bool term();
-    static LoadModule* loader(const char* filename, bool* invalid);
-    static LoadModule* loader(const char* data, uint32_t size, const char* mimeType, const char* rpath, bool copy);
-    static LoadModule* loader(const uint32_t* data, uint32_t w, uint32_t h, ColorSpace cs, bool copy);
-    static LoadModule* loader(const char* name, const char* data, uint32_t size, const char* mimeType, bool copy);
-    static LoadModule* font(const char* name);
-    static LoadModule* anyfont();
-    static bool retrieve(const char* filename);
-    static bool retrieve(LoadModule* loader);
+    std::function<bool(Paint* paint, const char* src, void* data)> func;
+    void* data;
+};
+
+struct Loader
+{
+    INLIST_ITEM(Loader);
+
+    //Use either hashkey(data) or hashpath(path)
+    uintptr_t hashkey = 0;
+    char* hashpath = nullptr;
+
+    FileType type;                                  //current loader file type
+    atomic<uint16_t> sharing{};                     //reference count
+    bool readied = false;                           //read done already.
+    bool cached = false;                            //cached for sharing
+
+    Loader(FileType type) : type(type) {}
+    virtual ~Loader()
+    {
+        tvg::free(hashpath);
+    }
+
+    void cache(uintptr_t data)
+    {
+        hashkey = data;
+        cached = true;
+    }
+
+    void cache(char* data)
+    {
+        hashpath = data;
+        cached = true;
+    }
+
+    virtual bool open(const char* path) { return false; }
+    virtual bool open(const char* data, uint32_t size, const char* rpath, bool copy) { return false; }
+    virtual bool resize(Paint* paint, float w, float h) { return false; }
+    virtual void sync() {};  //finish immediately if any async update jobs.
+
+    virtual bool read()
+    {
+        if (readied) return false;
+        readied = true;
+        return true;
+    }
+
+    virtual bool close()
+    {
+        if (sharing == 0) return true;
+        --sharing;
+        return false;
+    }
+
+    char* open(const char* path, uint32_t& size, bool text = false)
+    {
+#ifdef THORVG_FILE_IO_SUPPORT
+        auto f = fopen(path, text ? "r" : "rb");
+        if (!f) return nullptr;
+
+        fseek(f, 0, SEEK_END);
+
+        size = ftell(f);
+        if (size == 0) {
+            fclose(f);
+            return nullptr;
+        }
+
+        auto content = tvg::malloc<char>(sizeof(char) * (text ? size + 1 : size));
+        fseek(f, 0, SEEK_SET);
+        size = fread(content, sizeof(char), size, f);
+        if (text) content[size] = '\0';
+
+        fclose(f);
+
+        return content;
+#endif
+        return nullptr;
+    }
+};
+
+
+struct ImageLoader : Loader
+{
+    static atomic<ColorSpace> cs;                   //desired value
+
+    float w = 0, h = 0;                             //default image size
+    RenderSurface surface;
+
+    ImageLoader(FileType type) : Loader(type) {}
+
+    virtual bool animatable() { return false; }  //true if this loader supports animation.
+    virtual Paint* paint() { return nullptr; }
+    virtual void set(const AssetResolver* resolver) {}
+
+    virtual RenderSurface* bitmap()
+    {
+        if (surface.data) return &surface;
+        return nullptr;
+    }
+};
+
+
+struct FontMetrics
+{
+    Point size;  //text width, height
+    float scale;
+    Point align{}, box{}, spacing{1.0f, 1.0f};
+    float fontSize = 0.0f;
+    uint32_t lines = 1;      //line count
+    TextWrap wrap = TextWrap::None;
+
+    void *engine = nullptr;  //engine extension
+
+    ~FontMetrics()
+    {
+        tvg::free(engine);
+    }
+};
+
+
+struct FontLoader : Loader
+{
+    static constexpr const float DPI = 96.0f / 72.0f;   //dpi base?
+
+    char* name = nullptr;
+
+    FontLoader(FileType type) : Loader(type) {}
+
+    using Loader::read;
+
+    virtual bool get(FontMetrics& fm, char* text, uint32_t len, RenderPath& out) = 0;
+    virtual void transform(Paint* paint, FontMetrics& fm, float italicShear) = 0;
+    virtual void release(FontMetrics& fm) = 0;
+    virtual void metrics(const FontMetrics& fm, TextMetrics& out) = 0;
+    virtual bool metrics(const FontMetrics& fm, const char* ch, GlyphMetrics& out) = 0;
+    virtual void copy(const FontMetrics& in, FontMetrics& out) = 0;
 };
 
 }
