+++ src/renderer/tvgLoaderMgr.cpp
